+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/equip/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.action Action} implementation to equip a player with an item.
+ */
+package org.betonquest.betonquest.quest.action.equip;